Move config file to `/config`.
